Offloaded Quadtree generation to web worker
