+++ tsconfig.json
@@ -3,7 +3,7 @@
     "target": "ESNext",
     "useDefineForClassFields": true,
     "module": "ESNext",
-    "lib": ["ESNext", "DOM"],
+    "lib": ["ESNext", "DOM", "WebWorker", "ScriptHost"],
     "moduleResolution": "Node",
     "strict": true,
     "resolveJsonModule": true,
@@ -14,5 +14,7 @@
     "skipLibCheck": true,
     "allowJs": true
   },
-  "include": ["src", "public/concatenateBlobs.js"]
+  "include": ["src", 
+ // "src/animweb/helpers/QuadTree.worker.jsree.worker.js"
+],
 }
